Support Dask Array inputs to dictionary learning
As Dask Arrays support the behavior we designed `LazyDataset`s for, it
would be better to just support passing Dask Arrays through instead of
`LazyDataset`. This way we can still specify the slicing that we would
like to use in advance before actually reading the data. Thus still
being able to delay reading until the worker starts its computation. The
big benefit is that this will allow us to deprecate `LazyDataset`s and
drop them from our maintenance load in the future.

In pursuit of this goal, we make some minor tweaks to
`block_generate_dictionary_parallel`'s `calculate_block` function so as
to handle Dask Arrays and `LazyDataset`s. Basically this amounts to a
coercion to a NumPy array and forcing synchronized non-threaded
"computation" of the array (i.e. reading the data). The last part is
necessary to ensure that the Distributed cluster is not used to retrieve
the data, which could either result in oversubscription of the workers
(i.e. deadlock) or an inefficient retrieval of data by different
workers. Forcing the computation to be synchronized and non-threaded
guarantees that the data will be read on the same worker that requests
it; thus, avoiding the aforementioned issues. This allows Dask Arrays to
be used in place of `LazyDataset`s for dictionary learning while
continuing to allow `LazyDataset`s for now.