Use Dask to load data for dictionary learning
Changes the dictionary learning step in the workflow from using a
`LazyZarrDataset` to using a Dask Array. Given the tweaks already made
to `halo_block_generate_dictionary_parallel`, this should have the same
effect internally of specify a chunk of data for each worker to read and
having them load it right before computing the result.