Use Dask to load data for postprocessing
Changes the postprocessing step in the workflow from using a
`LazyZarrDataset` to using a Dask Array. Given the tweaks already made
to `block_postprocess_data_parallel`, this should have the same effect
internally of specify a chunk of data for each worker to read and having
them load it right before computing the result.