Use Zarr Arrays directly with viewers
As we now no longer zip up Zarr groups, using Zarr directories instead,
there is no need for an object like a `LazyZarrDataset` to be used with
the image viewers after each step in the workflow. This is true because
`DirectoryStore`s already open files on demand instead of keeping them
open. Hence there is no need to actually try and close them ourselves
afterwards, which is the only reason `LazyZarrDataset` is used here.
Using the viewers with Zarr `Array`s directly should simplify the
surrounding code and eliminate yet another use of the
`LazyZarrDataset`s.