Explicitly load the label image from Zarr
If we don't explicitly load the label image, `lblimg`, from Zarr, it
appears the comparison with `0` (i.e. `lblimg==0` ) evaluates to `False`
resulting in no mask. As a result the background that does not
correspond a label is covered with an unnecessary color layer (with some
transparency). This explicitly loads the label image data into a NumPy
Array. Thus the non-labeled background is fixed to be completely
transparent in the label image layer.