Keep readonly Zarr Groups always open
Now that Zarr Groups are kept as directories and not stored as Zip
files, there is no real need to close them later as there is not really
a file handle associated with the directory. Only individual chunk files
are opened only long enough to get the data out and are then closed. As
such there is no benefit to closing the Zarr Groups. However there is
some cost incurred when it comes to inspecting data later or
intermediate computations with Dask. Hence this switches to simply
always keeping Zarr Groups open. That way any Dask objects that base
themselves on such Zarr Groups remain valid after a computation has
finished.