Create registration's frame shape arrays locally
There is really no benefit in constructing the `frame_shape` and
`half_frame_shape` arrays as Dask Arrays initially. They should only be
a couple elements long and so can be trivially constructed on the
notebook and then shared. Thus we create both arrays locally with NumPy
and then promote them to Dask Arrays through `asarray`, which makes them
one big chunk each (as we intend).