Consolidate `dask_store_zarr` calls to `store`
Instead of calling `store` separately for each Dask Array that we intend
to store with `dask_store_zarr`, pass them all in bulk to one `store`
call. This way the different arrays being stored share some intermediate
portions of the graph any Dask optimizations can be applied to all
arrays at once. Further the intermediate results that are computed can
be reused for all Dask Arrays instead of being computed for each one.
While we didn't tend to store multiple Dask Arrays, it is still helpful
to have this optimization in the few cases we do.