Reimplement `dask_io_remove` to be non-blocking
Instead of trying to do the file/directory removal in place, move the
stuff to remove into a temporary directory created for this purpose.
Then construct the directory removal tree on this temporary directory
and submit it to the Dask cluster (if the executor is provided). As the
undesirable contents have been moved out of the way, there is nothing
blocking writing new stuff in its place. Thus the contents to remove can
easily be dealt with asynchronously as new contents are written in its
place.