Return the future from `dask_io_remove`
Instead of trying to create a progress bar for removal (which is largely
uninteresting to the user), simply return the future if the computation
was already submitted. That way they can block or create progress bars
as needed. Alternatively they can just ignore the removal and proceed to
later computations since the removal is a non-blocking step.