Optionally use dask_drmaa in startup_distributed
Instead of simply assuming `dask_drmaa` is always available, attempt to
import `dask_drmaa` within `startup_distributed`. If `dask_drmaa`
imports fine, try to configure a DRMAA-based cluster to use with
`distributed` just as before.

However if `dask_drmaa` does not import fine or if something goes wrong
trying to configure the DRMAA-based cluster, fallback to a
`LocalCluster`, which provides the same number of workers. This can be
handy when running on a local machine where a DRMAA compatible cluster
is not available. Alternatively this can be used on a single node of the
cluster.