Stop workers on LocalCluster or DRMAACluster
In `shutdown_distributed`, provide some logic to use the `stop_worker`
API of `LocalCluster` if possible to manage the shutdown of workers.
However if it is not possible (like with `DRMAACluster`), use
`stop_workers` as we have been before. Also make sure to take a copy of
the `list` of workers before operating on them so that things don't get
messed up trying to iterate over a `list` that is changing.