Force Dask's DRMAACluster to 1 thread per worker
When we startup the Dask `LocalCluster`, we make sure to set the number
of threads per worker to 1 as we want to only use 1 core per worker.
However the same is not currently done with Dask's `DRMAACluster`. This
fixes that issue by ensuring the `dask-worker` CLI gets a command line
option to specify 1 core per worker. As the number of cores per worker
affects how memory limits are estimated, we definitely want to make sure
Dask doesn't think the upper bound is the total memory on the node, but
only the amount of memory given per slot.