Provide optional kwargs in startup_distributed
Adds the option in `startup_distributed` to provide some additional
keyword arguments for the cluster constructor and for the `Client`
constructor. This should provide some needed flexibility for when the
startup of the Distributed cluster and Client need further customization
(e.g. prefacing the startup function with something). Will be helpful
when running the workflow under singularity.