Register frame inverses
Finds the inverses of each frames after adjusting them so that the
lowest value out of all of them is 1. This ensures that the values are
not only all positive definite, but there won't be any problematic
values like infinity. Also the range of the values after inverting
should be `(0, 1]`. Stores these to another variable for easy access.
Makes sure to clean them up at the end. Updates comments to make this
change to the template clearer.