Drop check for all workers to be shutdown
As both clusters have already ensured that all of their workers are
shutdown before getting to this `while` loop, there is no need for this
`sleep` recheck `while` loop. So just drop this loop.

Should add that with the DRMAA Cluster on Tornado 5, this `sleep` was
causing everything to hang. Interrupting the code revealed that all
workers had been shutdown. Not sure whether this is a Tornado 5 bug, an
issue with the DRMAA Cluster, or our own fault. In any event, having
DRMAA block our process until all workers are shutdown seems like the
right thing to do and fixes this issue.