Create `DistributedDirectoryStore`
Adds a Zarr store based off of `DirectoryStore` called
`DistributedDirectoryStore`, which handles file and directory removal in
an asynchronous manner. This allows space to be made in the Zarr store
for a new key immediately while the actual directory and file cleanup is
handled in parallel in the background with Dask's Distributed library.
Using this store with Zarr should allow us the convenient API that Zarr
provides for setting and removing key-value pairs without having to
worry about performance issues that have been encountered on slow
storage mediums historically as this strategy is already well optimized
for that goal. Everything else about `DistributedDirectoryStore` is the
same as `DirectoryStore`.