Use `fire_and_forget` with file removal task
When `dask_io_remove` is provided with a path to remove and an
`executor` (now commonly called Client in Distributed lingo), it will
submit the removal task to the `executor` and return a `Future`. However
if the `Future` is not held onto, Distributed will cancel this job to
avoid extra work as it assumes we don't care whether it completes any
more. As the path removal task is not particularly interesting, we
frequently drop the `Future` associated meaning that content is left
stranded instead of being removed as it should be.

To fix this, we need to ensure that this removal task still completes
even if we don't track its progress while still giving a `Future` should
the user decide to follow it. To manage this, we submit the task with
our `executor` and then pass the returned `Future` into the `executor`'s
`fire_and_forget` method. This ensures that the Client will hang onto
the `Future` for us until completion even if we don't. Thus guaranteeing
the path will be removed. Afterwards we return the `Future` to the user
per usual.