Use fire_and_forget on registration shifts
As the registration shifts are always needed whether another loop of
registration is required or registration is complete and we want to
store the result, send the shifts to `fire_and_forget`. While we should
always hold on to the `shifts` (and thus the relevant `Futures`), go
ahead and make sure they will be computed by sending them to
`fire_and_forget`. That way we can sure the `shifts` will be computed no
matter what.