Use Dask Arrays in viewers
To make it a bit easier to manage loading in pieces of large image data,
use Dask Arrays to do the loading. While Zarr can certainly make this
work, using Dask Arrays comes with the bonus of lazily chain operations
until they are needed. Also enables a nice level of parallelism to load
in the data requested rapidly. Preps are computational pipeline for more
directly interacting with Dask Array computations in viewers.