Track progress of each registration with heuristic
There is no need to send the heuristic, `avg_rel_dist`, to
`client.compute` as passing it to `client.persist` is sufficient for
triggering the computation. Also the heuristic already has `Future`s
relevant to its progress embedded within its Dask graph, which
`progress` can detect. So just track progress through the heuristic
directly and replace it with its value when it is ready.