Create the `DistributedArrayStore`
Makes a general Array store called the `DistributedArrayStore` to handle
an arbitrary `MutableMapping` object for storage. It performs operations
in a nearly asynchronous manner. Storage proceeds through Dask Array's
`store` function and returns `Future`s that point to where the data was
stored on disk. This allows easy tracking of the progress if desired.
Also as the Dask Array is returned with `Future`s included it can be
chained into new Dask operations immediately and also stored
asynchronously. As results are stored to disk, they can be retrieved in
subsequent sessions easily without needing to repeat them.

Deletion occurs in a nearly asynchronous manner. It cancels all
`Future`s related to the storing the value. This ensures that anything
related to the previously stored value is treated as expired.
Computations that use the old Dask Array will raise `CancelledError`
when requested. Blocks on the underlying `MutableMapping` used for
storage to delete the value. However if the underlying `MutableMapping`
does this deletion asynchronously or nearly asynchronously, as
`DistributedDirectoryStore` does, this will also be a fast operation
even while deletion occurs in the background.

Before storing each array a nonce is used to make repeated tokens in
Dask Arrays unique. This important both to workaround and protect
against issues in Dask/Distributed regarding repeated tokens. By doing
this, it is possible to store the same Dask Array repeatedly to a key
and still trigger the full computation and store step. This can be
useful if the user discovers the storage failed for whatever reason and
needs to repeat it. Alternatively can be useful if the user accidentally
deleted the result previously and would like to store it again. Without
this the storage would be skipped entirely in subsequent calls (though
canceling `Future`s helps with that). Also this avoids some subtle
issues that can rise up in Distributed around repeated keys.

Some of the provided methods are rewritten to be more efficient for our
use case. These include `__contains__`, `clear`, `pop`, `popitem`, and
`update`. In the case of `__contains__`, we have better information
available than what a generic `MutableMapping` has (hence why it reads
values). So we can just check the underlying store for an entry instead.
In the case of `clear`, the standard `MutableMapping` implementation is
a bit inefficient as it uses `popitem`, which reads each value into
memory. We can do a bit better by simply grabbing each key and deleting
them. This implementation was sent upstream to CPython. The cases of
`pop` and `popitem` are a little more complicated and so are explained
below.

Generally operations on the `DistributedArrayStore` are nearly
asynchronous (due to some small setup cost where the operation blocks).
However `pop` and `popitem` are synchronous and come with non-trivial
cost. The reason being popping a value would mean getting the value and
deleting it. However as explained before, deletion invalidates all
`Future`s associated with the value. So that would make the Dask Array
returned useless as it doesn't actually contain the relevant data.

To fix this problem, we persist the data so it can be loaded into
Distributed where it will live in the memory of the workers (possibly
spilled to disk if it is too large). As we can't actually delete the
data on disk until we are sure it is all loaded, we wait on the
`Future`s associated with the persisted Dask Array. However there is a
second problem, namely that the typical deletion strategy involves
canceling the `Future`s associated with the loaded value. This would
cancel any downstream `Future`s including those persisted. To fix this
issue we wait until the old results are stored and then create a fresh
Dask Array with the stored values persisted and wait until it is loaded.
Only once they complete do we trigger deletion of the value in the
`MutableMapping` and the cancellation of their `Future`s. As the
upstream implementation of `popitem` doesn't use `pop`, rewrite ours to
use `pop` for simplicity.

While the `DistributedArrayStore` is a `MutableMapping`, it requires the
ability to create an empty dataset. It assumes the `MutableMapping`
provided for storage implements a `create_dataset` method for creating
an array-like storage location with specified name, shape, dtype, and
chunking (along with other optional keyword arguments). Methods like
this are provided by Zarr's and h5py's `Group`s so are generally
reasonable to expect. However users are free to subclass
`DistributedArrayStore` and override the `_create_dataset` method if
they need to do something else here.