Bump default Jupyter Notebook timeout to 60s
Appears the tests are getting closer to the default Jupyter Notebook
timeout of 30s and sometimes going over. So bump the timeout default for
our tests to 60s, which should provide plenty enough time for the tests
to complete. This change will carry over nicely to CI and the Docker
image build.