Publish to Client in DistributedArrayStore
Instead of keeping an internal `dict` for `Array`s, publish them as
datasets with `DistributedArrayStore`'s `Client` instance. This should
free the `Array`s from any dependence on particular workers holding the
`Future`s allowing those workers to be taken offline if needed. Also
should allow users to get a hold of the results outside of the
`DistributedArrayStore` instance if needed. To simplify the rewrite use
the `MutableMapping` interface to `datasets` to make the publishing and
unpublishing process similar to what we had in place with the internal
`dict`.

Note: As trying to get values that may not be present or set values that
may be present is noisy, include checks to ensure an item is dealt with
based on the knowledge of it being there or not.