Provide option for Adaptive Distributed clusters
Add an option to `startup_distributed` to configure the particular
`Cluster` instance for adaptive scaling. This allows jobs to be launched
on demand based on computational or memory load. It also allows for jobs
to be culled when they are not being used. Should be a bit more friendly
with cluster resources while being a bit nicer in terms of cost.
Providing an empty `dict` or more arguments to `adaptive_kwargs` will
enable it. Alternatively provided no `adaptive_kwargs` argument or
setting it to `None` will disable it.