Use Adaptive scaling in the workflow
Enable adaptive scaling in the workflow with default arguments.
Initialize the number of workers on the cluster to zero as adaptive
scaling will add them as needed. Also drop the `num_workers` argument
from the workflow and drop the import of `set_num_workers` as these are
no longer needed with adaptive scaling enabled.