Move Distributed cluster arguments to new cell
Separates out the Distributed cluster's keyword arguments into a new
cell after setting all of the paths. This is broken out as we don't want
to reset these values after starting up the cluster, but we may need to
refresh the paths. So it is better to have this in a separate cell.