Set default minimum and maximum workers
Provide default minimum and maximum workers for the Distributed Adaptive
cluster. The minimum is simply zero. The maximum is based on a single
machine instance. In this case, bound the maximum by the number of CPU
cores less one for running the Jupyter Notebook and Distributed
Scheduler. Different choices may be better suited for cluster
deployments.