Persist image min in registration initialization
As the minimum of the image stack to register is a particularly
important value that needs to be computed and used by many different
downstream computations when initializing the image registration
algorithm. Persist the computation of the minimum along with other
downstream computations to guarantee it is computed once and readily
reusable by downstream computations. Make sure that the scheduler keeps
track of the associated `Future`s even though we drop the minimum value
locally to guarantee it still gets computed.