Remove Arrays in Groups before removing Groups
This is important for unpublishing data as there is no concept of Groups
in published datasets, only a concept of keys. So removing the full path
is necessary to unpublish data. Also technically the keys in the
Distributed Array store include the full path and not the Group. Even
though treating the Group as existent works, it is technically an
unsupported hack. So remove all the Arrays and then remove the Group
they belong to.