Handle creation/deletion of Groups through Zarr
Instead of trying to deal with Zarr Group through the Distributed Array
Store hack, deal with them through Zarr directly. This is a bit more
robust strategy than was used previously and is not as much of a hack.
Would be better if this wasn't needed at all. However this will do until
we can figure out a better solution.