Use frame rolling to generate registered data
Instead of going through and manually slicing each Dask Array and
concatenating them back together, use the frame rolling functions to
align each frame as needed. Ensuring the origin is held fixed. Then
determine the amount to trim off the far end and slice it off. This is
significantly faster to generate the graph for, has a much simpler
graph, the code involved is much simpler, and preserves the chunking of
the input frames used for registration. Matches the result one would
have gotten with the old technique.