Allow Distributed Dashboard access from other IPs
Instead of restricting access to the Distributed Dashboard to the local
IP address, allow access from other IP addresses as well. This is needed
in containerized environments where there is network isolation and IP
addresses may not match across or outside of containers. Further this is
useful on HPCs where networking may be more complex to figure out and so
access needs to be a bit more flexible. That said, users are encouraged
to set specific IP address when they know them (particularly in less
secure use cases) in order to have at least some protection against
others accessing the Distributed instance.