Switch to `dask.config.set`
As `dask.set_options` is deprecated and `dask.config.set` is the
recommended alternative, just go ahead and switch. We are not too
interested in having backcompat with Dask in the workflow. After all we
are often adding features that we absolutely need here as soon as they
are released. So there is no point in being flexible on Dask version.