Drop key salting workaround
As `da.store` now uses UUIDs for all stored values in the Dask graph as
of Dask 0.18.0, there is no need to do that ourselves. So drop the hacky
key salting routine from our code. This also resolves a deprecation
warning issued by `map_blocks` due to the use of `token`, which also
appears in Dask 0.18.0.