Fix `fire_and_forget` usage bug
Even though we called `persist` on something that we want to reuse (i.e.
store), we didn't actually hold onto the object, but instead passed it
onto `fire_and_forget`. Then used the non-persisted value to store.
However this doesn't work as it appears `fire_and_forget` did not run it
after all.

To fix this issue, store all results we want to compute both before we
call `persist` and after. This way we construct the graph for things
like `min` once. Also we make sure we are holding on to all values we
called `persist` on after. By doing this, we can be sure those `Future`s
remain tracked until we are done with the computation.

After passing the final result off to be stored, go ahead and delete all
of the intermediate values. Distributed already knows about them at this
point and has added `Future`s into the graph for them that the storage
operation depends on. Hence we can be sure that they will all complete
and we don't have to worry about hanging on to them manually any more.