Drop IPython 3 support for the workflow
We are now long past IPython 3 and the big split. IPython 5 and friends
are the last release supporting Python 2 and are already well developed.
IPython 6 is out for Python 3+ and has seen a few releases. It's safe to
say that IPython 3 does not need to be considered here. So drop the code
designed to support the old monolithic IPython.