Drop `fire_and_forget` of registration shifts
In all cases we reuse the registration shifts we found. Either we use
them to measure how well we are doing between subsequent image
registration runs or we use them to apply the shifts to the images. Not
to mention we store the shifts right alongside the registered images.
Given all of this, it doesn't make sense to pass the shifts to
`fire_and_forget`. Not to mention the shifts are a mathematical
computation we use and not a task running in the background that we are
uninterested in following. Due to all of this, we drop the
`fire_and_forget` line with the image registration shifts.