Require `nbserverproxy` for Python 3.5+
To aid in making the Dask Distributed Bokeh-based Dashboard more
accessible (particularly remotely) require `nbserverproxy` on Python
versions that support it. This way the Dashboard will be proxied by the
Jupyter notebook allowing it to be accessed through the same port. Thus
avoiding the need to expose more ports with containers or setup more SSH
tunnels for remote usage.