Lower priority of persisting rechunked X
While we do want to ensure that the data being fit gets loaded in
advance to ensure it is easy to randomize for feeding into dictionary
learning, it's lower priority than any dictionary learning step (unless
it is needed for a running dictionary learning step). So drop the
priority of this task by a floating point epsilon below the default
priority. This way it remains close to the default priority and still
gets started earlier, but will nice itself if dictionary learning is not
reliant on the particular values being served.