Eliminate persisted transpose of raw data
When finding the sparse code previously we computed the covariance in
the expected way. However as this relied on transposing the raw data, it
resulted in a copy of the raw data transposed remaining in persisted
data long before reaching the sparse code computation. To fix this,
rewrite the covariance computation to only rely on the raw data and not
its transposed form. This results in an extra transpose being performed
on the final result, but is tolerable compared to holding the raw data
in transposed form in distributed memory. Thus we are able to pull the
raw data from disk once we are ready for that computation only
temporarily using this memory. Hopefully avoids keeping too many workers
occupied holding that data as well.