Get gramian and covariance matrices with tensordot
Instead of using `da.dot` and transposing the matrices, rewrite the Dask
Array computations for the gramian and covariance matrices to just use
`da.tensordot`. While this does look a little uglier code wise, this
streamlines the Dask graph a bit and should extra computation/memory
needed to find the transpose of the dictionary as well.