Compute sparse code on cluster after
Instead of computing the dictionary and sparse code in the same worker,
switch back to the regular cluster for computing the sparse code. For
dictionary learning, it is very valuable to compute everything on a
single worker and thus benefit from data locality and using BLAS with
multiple cores as these are very important especially when computing the
dictionary over the full dataset. However the sparse code only needs to
look at some piece of the data it needs to encode with the dictionary.
Hence these can be run in an embarrassingly parallel fashion. So this
benefits more from having access to the cluster then dictionary learning
itself. As such we make use of the regular cluster for this step once
dictionary learning is complete.