Persist dictionary learning results at end of loop
Instead of persisting data going into dictionary learning, which we are
not particularly interested in holding onto at least when initializing
dictionary learning, persist results from dictionary learning at the end
of each iteration. This persists only the necessary values that will be
needed for the next pass. Thus it avoids persisting things like the next
batch of randomized data to learn with as it can be generated at any
time. Further it avoids persisting initialization data, which is either
trivial, already available, or can be quickly made available for the
computation when it starts.