Set OpenBLAS threads after importing Distributed
For some reason, setting the number of threads for OpenBLAS to use first
does not work reliably on the cluster. So import `distributed.joblib`
first and then set the number of threads for OpenBLAS to use. This
should give the worker enough time to get setup beforehand. Seems to
work in practice on the cluster.