Use new Bokeh `js_on_change` event hook
In Bokeh 0.12.4, the `js_on_change` method was added for hooking event
callbacks onto any `Model` object. In Bokeh 0.12.15, new selection model
was added, which allowed use of the `js_on_change` method. As of Bokeh
1.0.0, `selected.js_on_change` is the preferred way of managing custom
callback events for selections and the old `callback` strategy used here
is dropped. So this makes some needed code changes to adapt to the
`selected.js_on_change` strategy.