Slowly peel away `list` wrappers from `atop` call
The code here was unintentionally hard-coded to work with 2-D frames. In
reality, we would rather work with N-D frames. This means we have to
remove some number of `list` wrappers from our array that were
introduced by `atop`. So update the code to do that iteratively until
all `list` wrappers are removed.