Persist mask of best FFT overlaps
This appears to be crucial to avoid a bug either introduced or uncovered
by Dask 0.20.0's subgraph callable atop optimizations. Plus as the
overlap is needed to find the appropriate shifts for each frame, go
ahead and persist it along with its intermediate values to compute this
part of the graph. All intermediate values are immediately freed to
ensure that persisting them doesn't result in any additional long term
storage overhead.