Require Bokeh 1.0.0+
As there appears to be a bug in earlier versions of Bokeh's `Selection`
model, which resulted in the model being replaced after each event,
selections after the first one don't trigger the callback. This breaks
the ROI trace plot as one can only select the first ROI and then things
stop working. Fortunately this issue was fixed in Bokeh 1.0.0. Though
the details escape me. Given this, go ahead and bump the minimum
requirement to Bokeh 1.0.0. As a 1.0.0 release is now available, this
makes sense to have as a minimum version constraint anyways.