Create shuffled dictionary seed data
Create a separate step that handles shuffling the initial data and
randomly subsampling frames to seed the initial dictionary. This allows
for the random sampling to happen in parallel. Also allows dictionary
learning to happen under different conditions where it need not be
concerned about the performance of shuffling the data.