Use Subresource Integrity with Pako
To ensure the CDN provided us with the source we expect for Pako's
`pako_inflate.min.js`, use Subresource Integrity to specify a hash for
the Pako resource. This hash was generated using a cloned copy of the
GitHub repo, checking out the 1.0.4 tag, and hashing the same prebuilt
resource that was checked in there.