Generate hashes for Bokeh resources
Adds some utility functions to generates hashes for Bokeh resources
based on our local copies of these resources installed with the Conda
packages. Thus we are able to generate them dynamically for different
versions of Bokeh while still basing them off of resource files we
trust. Includes the hashes in the HTML document so as to use Subresource
Integrity to ensure the expected sources are gotten from the CDN.