Inject input data's keys directly into the graph
Makes direct usage of each chunk key in the input data without making
explicit use of slicing. Should cut a few intermediate steps out a few
needless steps in the graph. Also should make it easier for Dask to
realize that it has this data basically ready to go when needed.
Simplifies the code a bit as well.