Inline inner stats tuple handling
Previously we made `tuple` packing and unpacking for `inner_stats` part
of the Dask graph. However this creates some largely needless tasks for
the purpose of playing nicely with how scikit-learn chooses to group the
data. As we already wrap scikit-learn's `dict_learning_online` function,
we can actually handle `tuple` packing and unpacking within this wrapper
function. Thus we update the wrapper function to take in the statistics
as two separate arguments and return a flat tuple with the components
and both statistics at the same level. This simplifies the Dask graph
generated a bit and may help a little bit with performance.