Reference stored data in `components_`
As the `dictionary` is already computed and available through our stored
copy, there is no need to hold any references to the pre-stored copy as
they can all reference the stored version. Thus go ahead and replace the
`learner`'s `components_` with a reshaped version of the stored data.
By freeing all references to the pre-stored copy of the dictionary, Dask
is able to free this persisted memory as soon as it is done with it.
Thus keeping resource usage light.